Move allauth forms to adapter.py (#1145)
